bindings: move bindings
